requirements.txt
PyYAML
flask
docker